progress

+++ common/inc/comboot.h
@@ -8,7 +8,7 @@ struct com_boot_info {
         struct com_gfx_frame_buf frame_buf;
         struct com_gfx_psf_font font;
         struct com_mem_map mem_map;
-        int page_size;
+        size_t page_size;
 };
 
 #endif
